backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.0
uvicorn[standard]==0.24.0
pydantic==2.4.0
pydantic-settings==2.0.0

# Database
sqlalchemy==2.0.0
alembic==1.12.0
redis==5.0.0

# AI and ML
google-generativeai==0.3.0
openai==1.3.0
sentence-transformers==2.2.0
qdrant-client==1.6.0

# Web scraping and data collection
requests==2.31.0
beautifulsoup4==4.12.0
scrapy==2.11.0
feedparser==6.0.10
PyGithub==1.59.0
aiohttp==3.9.0

# Image generation
pillow==10.0.0
requests-oauthlib==1.3.1

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.25.0
aiofiles==23.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development and testing
pytest==7.4.0
pytest-asyncio==0.21.0
black==22.12.0
isort==5.12.0
flake8==6.0.0
mypy==1.6.0

# Monitoring and logging
structlog==23.1.0
sentry-sdk[fastapi]==1.38.0
